moves enums
