+++ lib/due_date.dart
@@ -24,12 +24,12 @@
 library due_date;
 
 import 'src/due_date.dart';
-import 'src/enums.dart' show WeekdayOccurrence;
+import 'src/enums/enums.dart' show WeekdayOccurrence;
 import 'src/everies/everies.dart';
 
 export 'src/date_validators/date_validators.dart';
 export 'src/due_date.dart';
-export 'src/enums.dart' show Month, Week, Weekday, WeekdayOccurrence;
+export 'src/enums/enums.dart' show Month, Week, Weekday, WeekdayOccurrence;
 export 'src/everies/everies.dart';
 export 'src/extensions.dart'
     show
