+++ lib/src/everies/every_weekday.dart
@@ -1,7 +1,7 @@
 import 'package:time/time.dart';
 
 import '../date_validators/date_validators.dart';
-import '../enums.dart';
+import '../enums/enums.dart';
 import 'every_date_validator.dart';
 import 'every_date_validator_union.dart';
 import 'every_month.dart';
