+++ test/enums/week_test.dart
@@ -1,225 +1,8 @@
-import 'package:due_date/due_date.dart';
 import 'package:due_date/period.dart';
 import 'package:test/test.dart';
 import 'package:time/time.dart';
 
 void main() {
-  group('Weekday:', () {
-    group('Throw on factory outside of range:', () {
-      test('Value below 1', () {
-        expect(() => Weekday.fromDateTimeValue(0), throwsRangeError);
-      });
-      test('Value above 7', () {
-        expect(() => Weekday.fromDateTimeValue(8), throwsRangeError);
-      });
-    });
-    group('Previous:', () {
-      for (final weekday in Weekday.values) {
-        test(weekday.name, () {
-          if (weekday != Weekday.monday) {
-            expect(
-              weekday.previous,
-              equals(Weekday.fromDateTimeValue(weekday.dateTimeValue - 1)),
-            );
-          } else {
-            expect(weekday.previous, equals(Weekday.sunday));
-          }
-        });
-      }
-    });
-    group('Next:', () {
-      for (final weekday in Weekday.values) {
-        test(weekday.name, () {
-          if (weekday != Weekday.sunday) {
-            expect(
-              weekday.next,
-              equals(Weekday.fromDateTimeValue(weekday.dateTimeValue + 1)),
-            );
-          } else {
-            expect(weekday.next, equals(Weekday.monday));
-          }
-        });
-      }
-    });
-    group('WeekendDays:', () {
-      final set = {Weekday.saturday, Weekday.sunday};
-      test('Contains $set', () {
-        expect(
-          Weekday.weekend,
-          containsAllInOrder(set),
-        );
-      });
-      test('Is ${set.runtimeType}', () {
-        expect(
-          Weekday.weekend,
-          isA<Set<Weekday>>(),
-        );
-      });
-    });
-    group('Workdays:', () {
-      final set = {
-        Weekday.monday,
-        Weekday.tuesday,
-        Weekday.wednesday,
-        Weekday.thursday,
-        Weekday.friday,
-      };
-      test('Contains $set', () {
-        expect(
-          Weekday.workdays,
-          containsAllInOrder(set),
-        );
-      });
-      test('Is ${set.runtimeType}', () {
-        expect(
-          Weekday.workdays,
-          isA<Set<Weekday>>(),
-        );
-      });
-    });
-    group('fromThisWeek:', () {
-      final date = DateTime.utc(2022, DateTime.august, 8);
-      for (final weekday in Weekday.values) {
-        test(weekday.name, () {
-          expect(
-            weekday.fromWeekOf(date),
-            equals(DateTime.utc(2022, DateTime.august, 8 + weekday.index)),
-          );
-        });
-      }
-    });
-    group('occurrencesIn august 2022', () {
-      final month = DateTime.utc(2022, DateTime.august);
-      test('monday', () {
-        const monday = Weekday.monday;
-        expect(
-          monday.occurrencesIn(month.year, month.month),
-          equals(5),
-        );
-      });
-      test('tuesday', () {
-        const tuesday = Weekday.tuesday;
-        expect(
-          tuesday.occurrencesIn(month.year, month.month),
-          equals(5),
-        );
-      });
-      test('wednesday', () {
-        const wednesday = Weekday.wednesday;
-        expect(
-          wednesday.occurrencesIn(month.year, month.month),
-          equals(5),
-        );
-      });
-      test('thursday', () {
-        const thursday = Weekday.thursday;
-        expect(
-          thursday.occurrencesIn(month.year, month.month),
-          equals(4),
-        );
-      });
-      test('friday', () {
-        const friday = Weekday.friday;
-        expect(
-          friday.occurrencesIn(month.year, month.month),
-          equals(4),
-        );
-      });
-      test('saturday', () {
-        const saturday = Weekday.saturday;
-        expect(
-          saturday.occurrencesIn(month.year, month.month),
-          equals(4),
-        );
-      });
-      test('sunday', () {
-        const sunday = Weekday.sunday;
-        expect(
-          sunday.occurrencesIn(month.year, month.month),
-          equals(4),
-        );
-      });
-    });
-  });
-  group('Month:', () {
-    group('Throw on factory outside of range:', () {
-      test('Value below 1', () {
-        expect(() => Month.fromDateTimeValue(0), throwsRangeError);
-      });
-      test('Value above 12', () {
-        expect(() => Month.fromDateTimeValue(13), throwsRangeError);
-      });
-    });
-    group('Previous:', () {
-      for (final month in Month.values) {
-        test(month.name, () {
-          if (month != Month.january) {
-            expect(
-              month.previous,
-              equals(Month.fromDateTimeValue(month.dateTimeValue - 1)),
-            );
-          } else {
-            expect(month.previous, equals(Month.december));
-          }
-        });
-      }
-    });
-    group('Next:', () {
-      for (final month in Month.values) {
-        test(month.name, () {
-          if (month != Month.december) {
-            expect(
-              month.next,
-              equals(Month.fromDateTimeValue(month.dateTimeValue + 1)),
-            );
-          } else {
-            expect(month.next, equals(Month.january));
-          }
-        });
-      }
-    });
-    group('from', () {
-      const monthGenerator = MonthGenerator();
-      test('august', () {
-        final august = monthGenerator.of(DateTime(2022, DateTime.august));
-        expect(Month.august.of(2022, utc: false), equals(august));
-      });
-      test('august utc', () {
-        final august = monthGenerator.of(DateTime.utc(2022, DateTime.august));
-        expect(Month.august.of(2022), equals(august));
-      });
-      test('september', () {
-        final september = monthGenerator.of(DateTime(2022, DateTime.september));
-        expect(Month.september.of(2022, utc: false), equals(september));
-      });
-      test('september utc', () {
-        final september = monthGenerator.of(
-          DateTime.utc(2022, DateTime.september),
-        );
-        expect(Month.september.of(2022), equals(september));
-      });
-      test('february 2022', () {
-        final february = monthGenerator.of(DateTime(2022, DateTime.february));
-        expect(Month.february.of(2022, utc: false), equals(february));
-      });
-      test('february 2022 utc', () {
-        final february = monthGenerator.of(
-          DateTime.utc(2022, DateTime.february),
-        );
-        expect(Month.february.of(2022), equals(february));
-      });
-      test('february 2020', () {
-        final february = monthGenerator.of(DateTime(2020, DateTime.february));
-        expect(Month.february.of(2020, utc: false), equals(february));
-      });
-      test('february 2020 utc', () {
-        final february = monthGenerator.of(
-          DateTime.utc(2020, DateTime.february),
-        );
-        expect(Month.february.of(2020), equals(february));
-      });
-    });
-  });
   group('Week', () {
     group('Week.from', () {
       const year = 2022;
@@ -600,21 +383,6 @@ void main() {
             );
           });
         }
-      });
-    });
-  });
-  group('WeekdayOccurrence', () {
-    group('Equals', () {
-      test('FirstMonday', () {
-        const weekdayOccurrence = WeekdayOccurrence.firstMonday;
-        const everyWeekdayCountInMonth = EveryWeekdayCountInMonth(
-          week: Week.first,
-          day: Weekday.monday,
-        );
-        expect(
-          weekdayOccurrence,
-          equals(everyWeekdayCountInMonth),
-        );
       });
     });
   });
